--Added new table events_attendees, added related methods in model, controller and in test. Also changed a few migration files(first check if table exits)

+++ app/database/migrations/2014_02_04_205944_drop_ingredients_recipeingredients__table.php
@@ -12,7 +12,9 @@ class DropIngredientsRecipeingredientsTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::drop('ingredients_recipeingredients');
+		if( Schema::hasTable('ingredients_recipeingredients')){
+			Schema::drop('ingredients_recipeingredients');
+		}		
 	}
 
 	/**
