+++ app/database/migrations/2014_02_04_210000_drop_meals_tags_table.php
@@ -12,7 +12,9 @@ class DropMealsTagsTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::drop('meals_tags');
+		if(Schema::hasTable('meals_tags')){
+			Schema::drop('meals_tags');
+		}		
 	}
 
 	/**
